// React/Site/src/API.ts
export type ModelStringInput = {
  ne?: string | null,
  eq?: string | null,
  le?: string | null,
  lt?: string | null,
  ge?: string | null,
  gt?: string | null,
  contains?: string | null,
  notContains?: string | null,
  between?: Array< string | null > | null,
  beginsWith?: string | null,
  attributeExists?: boolean | null,
  attributeType?: ModelAttributeTypes | null,
  size?: ModelSizeInput | null,
};

export enum ModelAttributeTypes {
  binary = "binary",
  binarySet = "binarySet",
  bool = "bool",
  list = "list",
  map = "map",
  number = "number",
  numberSet = "numberSet",
  string = "string",
  stringSet = "stringSet",
  _null = "_null",
}


export type ModelSizeInput = {
  ne?: number | null,
  eq?: number | null,
  le?: number | null,
  lt?: number | null,
  ge?: number | null,
  gt?: number | null,
  between?: Array< number | null > | null,
};

export type ModelIDInput = {
  ne?: string | null,
  eq?: string | null,
  le?: string | null,
  lt?: string | null,
  ge?: string | null,
  gt?: string | null,
  contains?: string | null,
  notContains?: string | null,
  between?: Array< string | null > | null,
  beginsWith?: string | null,
  attributeExists?: boolean | null,
  attributeType?: ModelAttributeTypes | null,
  size?: ModelSizeInput | null,
};

export type ModelIntInput = {
  ne?: number | null,
  eq?: number | null,
  le?: number | null,
  lt?: number | null,
  ge?: number | null,
  gt?: number | null,
  between?: Array< number | null > | null,
  attributeExists?: boolean | null,
  attributeType?: ModelAttributeTypes | null,
};

export type User = {
  __typename: "User",
  id: string,
  fullName: string,
  firebaseUid: string,
  email: string,
  displayName: string,
  imageUri?: string | null,
  defaultMidiDevice?: string | null,
  purchasedTokens: number,
  remainingTokens: number,
  totalStreams: number,
  totalSeconds: number,
  lastStream?: string | null,
  createdInstanceId?: string | null,
  hostRoomUser?: ModelHostRoomUserConnection | null,
  createdAt: string,
  updatedAt: string,
};

export type ModelHostRoomUserConnection = {
  __typename: "ModelHostRoomUserConnection",
  items:  Array<HostRoomUser | null >,
  nextToken?: string | null,
};

export type HostRoomUser = {
  __typename: "HostRoomUser",
  id: string,
  userId: string,
  hostRoomId: string,
  lastSeen?: string | null,
  archived: boolean,
  pinned: boolean,
  user?: User | null,
  hostRoom?: HostRoom | null,
  createdAt: string,
  updatedAt: string,
  userHostRoomUserId?: string | null,
  hostRoomHostRoomUsersId?: string | null,
  hostRoomUserUserId?: string | null,
};

export type HostRoom = {
  __typename: "HostRoom",
  id: string,
  name: string,
  desc?: string | null,
  imageUri?: string | null,
  createdUserId: string,
  lastMessageId?: string | null,
  lastMessage?: HostRoomChatMessage | null,
  chatMessages?: ModelHostRoomChatMessageConnection | null,
  hostRoomUsers?: ModelHostRoomUserConnection | null,
  createdAt: string,
  updatedAt: string,
  hostRoomLastMessageId?: string | null,
};

export type HostRoomChatMessage = {
  __typename: "HostRoomChatMessage",
  id: string,
  userId: string,
  roomId: string,
  message: string,
  read: boolean,
  user?: User | null,
  hostRoom?: HostRoom | null,
  createdAt: string,
  updatedAt: string,
  hostRoomChatMessagesId?: string | null,
  hostRoomChatMessageUserId?: string | null,
  hostRoomChatMessageHostRoomId?: string | null,
};

export type ModelHostRoomChatMessageConnection = {
  __typename: "ModelHostRoomChatMessageConnection",
  items:  Array<HostRoomChatMessage | null >,
  nextToken?: string | null,
};
// ...
export type ModelBooleanInput = {
  ne?: boolean | null,
  eq?: boolean | null,
  attributeExists?: boolean | null,
  attributeType?: ModelAttributeTypes | null,
};
// ...
export type CreateHostRoomChatMessageInput = {
  id?: string | null,
  userId: string,
  roomId: string,
  message: string,
  read: boolean,
  hostRoomChatMessagesId?: string | null,
  hostRoomChatMessageUserId?: string | null,
  hostRoomChatMessageHostRoomId?: string | null,
};
// ...
export type DeleteHostRoomChatMessageInput = {
  id: string,
};
// ...
export type ModelHostRoomChatMessageFilterInput = {
  id?: ModelIDInput | null,
  userId?: ModelIDInput | null,
  roomId?: ModelIDInput | null,
  message?: ModelStringInput | null,
  read?: ModelBooleanInput | null,
  and?: Array< ModelHostRoomChatMessageFilterInput | null > | null,
  or?: Array< ModelHostRoomChatMessageFilterInput | null > | null,
  not?: ModelHostRoomChatMessageFilterInput | null,
  hostRoomChatMessagesId?: ModelIDInput | null,
  hostRoomChatMessageUserId?: ModelIDInput | null,
  hostRoomChatMessageHostRoomId?: ModelIDInput | null,
};
// ...
export type CreateHostRoomChatMessageMutation = {
  createHostRoomChatMessage?:  {
    __typename: "HostRoomChatMessage",
    id: string,
    userId: string,
    roomId: string,
    message: string,
    read: boolean,
    user?:  {
      __typename: "User",
      id: string,
      fullName: string,
      firebaseUid: string,
      email: string,
      displayName: string,
      imageUri?: string | null,
      defaultMidiDevice?: string | null,
      purchasedTokens: number,
      remainingTokens: number,
      totalStreams: number,
      totalSeconds: number,
      lastStream?: string | null,
      createdInstanceId?: string | null,
      hostRoomUser?:  {
        __typename: "ModelHostRoomUserConnection",
        nextToken?: string | null,
      } | null,
      createdAt: string,
      updatedAt: string,
    } | null,
    hostRoom?:  {
      __typename: "HostRoom",
      id: string,
      name: string,
      desc?: string | null,
      imageUri?: string | null,
      createdUserId: string,
      lastMessageId?: string | null,
      lastMessage?:  {
        __typename: "HostRoomChatMessage",
        id: string,
        userId: string,
        roomId: string,
        message: string,
        read: boolean,
        createdAt: string,
        updatedAt: string,
        hostRoomChatMessagesId?: string | null,
        hostRoomChatMessageUserId?: string | null,
        hostRoomChatMessageHostRoomId?: string | null,
      } | null,
      chatMessages?:  {
        __typename: "ModelHostRoomChatMessageConnection",
        nextToken?: string | null,
      } | null,
      hostRoomUsers?:  {
        __typename: "ModelHostRoomUserConnection",
        nextToken?: string | null,
      } | null,
      createdAt: string,
      updatedAt: string,
      hostRoomLastMessageId?: string | null,
    } | null,
    createdAt: string,
    updatedAt: string,
    hostRoomChatMessagesId?: string | null,
    hostRoomChatMessageUserId?: string | null,
    hostRoomChatMessageHostRoomId?: string | null,
  } | null,
};
// ...
export type DeleteHostRoomChatMessageMutation = {
  deleteHostRoomChatMessage?:  {
    __typename: "HostRoomChatMessage",
    id: string,
    userId: string,
    roomId: string,
    message: string,
    read: boolean,
    user?:  {
      __typename: "User",
      id: string,
      fullName: string,
      firebaseUid: string,
      email: string,
      displayName: string,
      imageUri?: string | null,
      defaultMidiDevice?: string | null,
      purchasedTokens: number,
      remainingTokens: number,
      totalStreams: number,
      totalSeconds: number,
      lastStream?: string | null,
      createdInstanceId?: string | null,
      hostRoomUser?:  {
        __typename: "ModelHostRoomUserConnection",
        nextToken?: string | null,
      } | null,
      createdAt: string,
      updatedAt: string,
    } | null,
    hostRoom?:  {
      __typename: "HostRoom",
      id: string,
      name: string,
      desc?: string | null,
      imageUri?: string | null,
      createdUserId: string,
      lastMessageId?: string | null,
      lastMessage?:  {
        __typename: "HostRoomChatMessage",
        id: string,
        userId: string,
        roomId: string,
        message: string,
        read: boolean,
        createdAt: string,
        updatedAt: string,
        hostRoomChatMessagesId?: string | null,
        hostRoomChatMessageUserId?: string | null,
        hostRoomChatMessageHostRoomId?: string | null,
      } | null,
      chatMessages?:  {
        __typename: "ModelHostRoomChatMessageConnection",
        nextToken?: string | null,
      } | null,
      hostRoomUsers?:  {
        __typename: "ModelHostRoomUserConnection",
        nextToken?: string | null,
      } | null,
      createdAt: string,
      updatedAt: string,
      hostRoomLastMessageId?: string | null,
    } | null,
    createdAt: string,
    updatedAt: string,
    hostRoomChatMessagesId?: string | null,
    hostRoomChatMessageUserId?: string | null,
    hostRoomChatMessageHostRoomId?: string | null,
  } | null,
};
// ...
export type ListHostRoomChatMessagesQuery = {
  listHostRoomChatMessages?:  {
    __typename: "ModelHostRoomChatMessageConnection",
    items:  Array< {
      __typename: "HostRoomChatMessage",
      id: string,
      userId: string,
      roomId: string,
      message: string,
      read: boolean,
      user?:  {
        __typename: "User",
        id: string,
        fullName: string,
        firebaseUid: string,
        email: string,
        displayName: string,
        imageUri?: string | null,
        defaultMidiDevice?: string | null,
        purchasedTokens: number,
        remainingTokens: number,
        totalStreams: number,
        totalSeconds: number,
        lastStream?: string | null,
        createdInstanceId?: string | null,
        createdAt: string,
        updatedAt: string,
      } | null,
      hostRoom?:  {
        __typename: "HostRoom",
        id: string,
        name: string,
        desc?: string | null,
        imageUri?: string | null,
        createdUserId: string,
        lastMessageId?: string | null,
        createdAt: string,
        updatedAt: string,
        hostRoomLastMessageId?: string | null,
      } | null,
      createdAt: string,
      updatedAt: string,
      hostRoomChatMessagesId?: string | null,
      hostRoomChatMessageUserId?: string | null,
      hostRoomChatMessageHostRoomId?: string | null,
    } | null >,
    nextToken?: string | null,
  } | null,
};